chore: rename org.dashj.sdk to org.dashj.platform.sdk for platform-core
