+++ examples/src/main/kotlin/org/dashj/platform/examples/DisplayContracts.kt
@@ -6,8 +6,8 @@
  */
 package org.dashj.platform.examples
 
-import org.dashevo.Client
-import org.dashevo.client.ClientOptions
+import org.dashj.platform.sdk.Client
+import org.dashj.platform.sdk.client.ClientOptions
 import org.json.JSONObject
 
 class DisplayContracts {
