+++ platform-core/src/main/kotlin/org/dashj/platform/sdk/client/ClientApps.kt
@@ -5,7 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
-package org.dashevo.client
+package org.dashj.platform.sdk.client
 
 class ClientApps(apps: Map<String, ClientAppDefinition>) {
     private val apps = hashMapOf<String, ClientAppDefinition>()
