+++ platform-core/src/main/kotlin/org/dashj/platform/sdk/client/ClientOptions.kt
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-package org.dashevo.client
+package org.dashj.platform.sdk.client
 
 import org.dashj.platform.dapiclient.DapiClient
 import org.dashj.platform.dapiclient.provider.DAPIAddressListProvider
