+++ platform-core/src/main/kotlin/org/dashj/platform/sdk/platform/Contracts.kt
@@ -4,17 +4,17 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-package org.dashevo.platform
+package org.dashj.platform.sdk.platform
 
+import java.util.Date
 import org.bitcoinj.core.ECKey
-import org.dashevo.client.ClientAppDefinition
 import org.dashj.platform.dpp.contract.DataContract
 import org.dashj.platform.dpp.contract.DataContractCreateTransition
 import org.dashj.platform.dpp.identifier.Identifier
 import org.dashj.platform.dpp.identity.Identity
+import org.dashj.platform.sdk.client.ClientAppDefinition
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import java.util.Date
 
 class Contracts(val platform: Platform) {
