+++ platform-core/src/main/kotlin/org/dashj/platform/sdk/platform/multicall/MulticallException.kt
@@ -1,3 +1,3 @@
-package org.dashevo.platform.multicall
+package org.dashj.platform.sdk.platform.multicall
 
 class MulticallException(val exceptionList: List<Exception>) : Exception("Multicall exception: $exceptionList")
