+++ platform-core/src/main/kotlin/org/dashj/platform/sdk/platform/multicall/MulticallMethod.kt
@@ -1,4 +1,4 @@
-package org.dashevo.platform.multicall
+package org.dashj.platform.sdk.platform.multicall
 
 interface MulticallMethod<T> {
     fun execute(): T
