+++ platform-core/src/main/kotlin/org/dashj/platform/sdk/platform/multicall/MulticallQuery.kt
@@ -4,7 +4,7 @@
  * This source code is licensed under the MIT license found in the
  * LICENSE file in the root directory of this source tree.
  */
-package org.dashevo.platform.multicall
+package org.dashj.platform.sdk.platform.multicall
 
 import org.slf4j.LoggerFactory
