+++ tools/src/main/java/org/dashj/platform/tools/WalletTool.java
@@ -132,9 +132,9 @@ import org.dashj.platform.dashpay.callback.RegisterPreorderCallback;
 import org.dashj.platform.dpp.document.Document;
 import org.dashj.platform.dpp.identifier.Identifier;
 import org.dashj.platform.dpp.identity.Identity;
-import org.dashevo.platform.DomainDocument;
-import org.dashevo.platform.Names;
-import org.dashevo.platform.Platform;
+import org.dashj.platform.sdk.platform.DomainDocument;
+import org.dashj.platform.sdk.platform.Names;
+import org.dashj.platform.sdk.platform.Platform;
 import org.jetbrains.annotations.NotNull;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
