+++ tools/src/main/kotlin/org/dashj/platform/tools/NetworkActivity.kt
@@ -6,16 +6,16 @@
  */
 package org.dashj.platform.tools
 
-import org.dashevo.Client
-import org.dashevo.client.ClientOptions
-import org.dashevo.platform.Documents
-import org.dashevo.platform.Names
-import org.dashevo.platform.Platform
 import org.dashj.platform.dapiclient.model.DocumentQuery
 import org.dashj.platform.dashpay.ContactRequest
 import org.dashj.platform.dashpay.ContactRequests
 import org.dashj.platform.dpp.document.Document
 import org.dashj.platform.dpp.identifier.Identifier
+import org.dashj.platform.sdk.Client
+import org.dashj.platform.sdk.client.ClientOptions
+import org.dashj.platform.sdk.platform.Documents
+import org.dashj.platform.sdk.platform.Names
+import org.dashj.platform.sdk.platform.Platform
 
 class NetworkActivity {
     companion object {
